+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 4.2 (unreleased)
 ================
 
+- Fixed #64 issue related to ``setup_class`` and ``fixture`` executions on rerun (Thanks to
+  `@OlegKuzovkov`_ for the PR).
 - Added new ``execution_count`` attribute to reflect the number of test case executions according to #67 issue.
   (Thanks to `@OlegKuzovkov`_ for the PR).